Create IRootShape, update converter and raise version number
