+++ Models/Functions/RootShape/RootShapeCylindre.cs
@@ -1,0 +1,61 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Models.Core;
+using Models.Interfaces;
+using APSIM.Shared.Utilities;
+using Models.PMF.Organs;
+using Models.Soils;
+
+namespace Models.Functions
+{
+    /// <summary>
+    /// This model calculates the proportion of each soil layer occupided by roots.
+    /// </summary>
+    [Serializable]
+    [ViewName("UserInterface.Views.GridView")]
+    [PresenterName("UserInterface.Presenters.PropertyPresenter")]
+    public class RootShapeCylindre : Model, IRootShape, ICustomDocumentation
+    {
+        /// <summary>Calculates the root area for a layer of soil</summary>
+        public void CalcRootProportionInLayers(ZoneState zone)
+        {
+            for (int layer = 0; layer < zone.soil.Thickness.Length; layer++)
+            {
+                double prop;
+                double top = layer == 0 ? 0 : MathUtilities.Sum(zone.soil.Thickness, 0, layer - 1);
+
+                if (zone.Depth < top)
+                {
+                    prop = 0;
+                } 
+                else
+                {
+                    prop = zone.soil.ProportionThroughLayer(layer, zone.Depth);
+                }
+                zone.RootProportions[layer] = prop;
+            }
+        }
+
+        /// <summary>Writes documentation for this function by adding to the list of documentation tags.</summary>
+        /// <param name="tags">The list of tags to add to.</param>
+        /// <param name="headingLevel">The level (e.g. H2) of the headings.</param>
+        /// <param name="indent">The level of indentation 1, 2, 3 etc.</param>
+        public void Document(List<AutoDocumentation.ITag> tags, int headingLevel, int indent)
+        {
+            if (IncludeInDocumentation)
+            {
+                // add a heading.
+                tags.Add(new AutoDocumentation.Heading(Name, headingLevel));
+
+                // add graph and table.
+                //tags.Add(new AutoDocumentation.Paragraph("<i>" + Name + " is calculated as a function of daily min and max temperatures, these are weighted toward VPD at max temperature according to the specified MaximumVPDWeight factor.  A value equal to 1.0 means it will use VPD at max temperature, a value of 0.5 means average VPD.</i>", indent));
+                //tags.Add(new AutoDocumentation.Paragraph("<i>MaximumVPDWeight = " + MaximumVPDWeight + "</i>", indent));
+
+                // write memos.
+                foreach (IModel memo in Apsim.Children(this, typeof(Memo)))
+                    AutoDocumentation.DocumentModel(memo, tags, headingLevel + 1, indent);
+            }
+        }
+    }
+}
